Address review feedback: extract FormScrollContext, use snapshotFlow
Committed-By-Agent: claude
